services/cart_service/pyproject.toml
[project]
name = "cart-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "click==8.3.0",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi-cli==0.0.14",
    "fastapi-cloud-cli==0.3.1",
    "fastapi[standard]==0.120.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "pika==1.3.2",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pygments==2.19.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "redis[asyncio]==7.0.1",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.1",
    "sentry-sdk==2.42.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.49.0",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]
